+++ src/services/users/dto/index.ts
@@ -1,1 +1,2 @@
+export * from './sign-in';
 export * from './sign-up';
