+++ src/services/users/dto/sign-up/post-dto.ts
@@ -60,16 +60,3 @@ export class SignUpBodyDto {
   @IsString()
   profileImage!: string;
 }
-
-export class SignInBodyDto {
-  @ApiProperty({ example: 'test@test.com', description: '이메일', required: true })
-  @IsNotEmpty()
-  @IsEmail()
-  email!: string;
-
-  @ApiProperty({ example: 'asdf1234!@', description: '패스워드', required: true })
-  @IsNotEmpty()
-  @IsString()
-  @MinLength(8, { message: 'Password required at least 8' })
-  password!: string;
-}
